[FIX] Impl and debug destroyBlocksAround() and using

+++ jni/include/GameProcessor.h
@@ -250,7 +250,7 @@ private:
   /// @param col Output column index of impacted block of current level.
   /// @return TRUE is some block has been impacted, FALSE if ball has left level's boundaries
   /// in order to avoid index of block out of bounds.
-  bool getImpactedBlock(GLfloat ball_x, GLfloat ball_y, size_t* row, size_t* col);
+  bool getImpactedBlock(GLfloat ball_x, GLfloat ball_y, int* row, int* col);
   /// @brief Corrects ball's visual position after collision and notifies
   /// rendering thread.
   /// @param new_x Corrected ball's center position along X axis.
